// fe-experiment/backend-for-ved/src/lib/enums/common.enums.ts
export enum Lang {
  RU = 'ru',
}

export enum TimestampType {
  BY_HOUR = 'by-hour',
  BY_DAY = 'by-day',
  BY_MONTH = 'by-month',
}

export enum AllCurrencies {
  RUB = 'rub',
  CNY = 'cny',
  JPY = 'jpy',
  USD = 'usd',
  EUR = 'eur',
  AED = 'aed',
  TRY = 'try',
  HKD = 'hkd',
  INR = 'inr',
  USDT = 'usdt',
  BTC = 'btc',
  ETH = 'eth',
  CAD = 'cad',
  SGD = 'sgd',
  GBP = 'gbp',
  THB = 'thb',
  CHF = 'chf',
}

export enum CurrencyDirection {
  UP = 'up',
  DOWN = 'down',
  NO = 'no',
}

export enum MimeTypes {
  JSON = 'application/json',
  PDF = 'application/pdf',
  ZIP = 'application/zip',
  DOCX = 'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
}